requirements.txt
# requirements.txt

pandas~=2.0.3
aspose-tasks
mesa~=3.0.3
JPype1